requirements.txt
# MoodSense — Requirements
# Python 3.11+ | macOS ARM64 (works on Linux for Streamlit Cloud)
# Install: pip install -r requirements.txt

# ── Core Data & ML ────────────────────────────────────────────────────
numpy>=1.24,<2.0
pandas>=2.0,<3.0
scikit-learn>=1.4,<2.0
scipy>=1.11,<2.0
xgboost>=2.0,<3.0
lightgbm>=4.0,<5.0
imbalanced-learn>=0.11,<1.0

# ── NLP ───────────────────────────────────────────────────────────────
sentence-transformers>=2.2,<4.0
transformers>=4.35,<5.0        # updated for torch.compiler compatibility
tokenizers>=0.13,<1.0
vaderSentiment>=3.3,<4.0
textblob>=0.17,<1.0

# ── PyTorch (fix torch.compiler error) ─────────────────────────────────
torch>=2.1.0,<3.0               # must be >=2.0 for torch.compiler
torchvision>=0.16.0,<1.0        # optional if using vision models
torchaudio>=2.1.0,<3.0          # optional if using audio models

# ── Vector Search ─────────────────────────────────────────────────────
faiss-cpu>=1.7,<2.0

# ── Experiment Tracking ───────────────────────────────────────────────
mlflow>=2.10,<3.0

# ── API Backend ───────────────────────────────────────────────────────
fastapi>=0.100,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.0,<3.0

# ── Streamlit App ─────────────────────────────────────────────────────
streamlit>=1.31,<2.0

# ── Visualization ─────────────────────────────────────────────────────
matplotlib>=3.7,<4.0
seaborn>=0.12,<1.0
plotly>=5.15,<6.0
wordcloud>=1.9,<2.0
shap>=0.42,<1.0

# ── Lyrics ────────────────────────────────────────────────
# lyrics.ovh is used at runtime via requests (no package needed)

# ── Utilities ─────────────────────────────────────────────────────────
python-dotenv>=1.0,<2.0
packaging>=24.2.0
protobuf>=6.31.1,<7.0.0
requests>=2.31,<3.0
tqdm>=4.65,<5.0
joblib>=1.3,<2.0
pillow>=10.0,<12.0

requests
faiss-cpu
huggingface_hub